feat(64bit): widen Layout/AttributeDescription counts + add View::element_count

+++ lib/include/public/mimir/view.hpp
@@ -26,13 +26,13 @@ enum class FormatKind { Float, Signed, Unsigned, SignedNormalized, UnsignedNorma
 struct Layout
 {
     // Layout size in cartesian (x,y,z) format.
-    unsigned int x = 0, y = 1, z = 1;
+    size_t x = 0, y = 1, z = 1;
 
     // Returns the total number of elements contained in this layout.
-    unsigned int getTotalCount() { return x * y * z; };
+    size_t getTotalCount() { return x * y * z; };
 
     // Creates a layout with default value of 1 for Y and Z axes.
-    static Layout make(unsigned int x, unsigned int y = 1, unsigned int z = 1)
+    static Layout make(size_t x, size_t y = 1, size_t z = 1)
     {
         return Layout{ .x = x, .y = y, .z = z };
     }
@@ -80,7 +80,7 @@ struct AttributeDescription
     AllocHandle source = {};
     // Number of elements contained in the source allocation.
     // Must be non-zero to describe a valid attribute.
-    unsigned int size = 0;
+    size_t size = 0;
     // Format description for the elements in the source array; ignored when sources is null.
     FormatDescription format = {};
     // Handle for the allocation containing indices referencing the source array.
